Second draft for now. Payment is done. My third push will update payment interactor.
