+++ src/main/java/PaymentInteractor.java
@@ -1,0 +1,2 @@
+public class PaymentInteractor {
+}
